backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
python-dotenv==1.0.1
pydantic[email]==2.9.2
